alioop-microservice-prototype/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
python-dotenv==1.2.1

# Messaging services (optional - comment out if not using)
sendgrid==6.11.0
twilio==9.3.7
